We'd like to move the token API out to a "token server" ("tokener", in keeping with the general "-er"/"-or" endings of `sdk/core`?) compartment.  This is one step towards supporting multiple separate allocators but is more proximally motivated by wanting to explore the use of software ocaps in the revoker API.

To avoid compounding cross-compartment calls in hot paths, we should make `token_obj_unseal` a library function, a la CheriOS's fast leaf cross-compartment game.  That means we need to...

- [x] Support multiple libraries, with different PCCs, in the compiler
- [x] Support privileged libraries, with constants in PCC, in the loader
- [X] Rewrite `token_obj_unseal` in assembler
- [x] Move `token_obj_unseal` into a new privileged library and expose it as an IRQs-deferred sentry
- [ ] Pass function pointers to
   - [x] `token_sealed_unsealed_alloc` (`malloc`)
   - [x] `token_sealed_alloc` (`malloc`)
   - [ ] `token_obj_destroy` (`free`)
   - [ ] `token_allocate` (`malloc`)
- [ ] Move those functions to the library
- [x] Make the allocator use `token_obj_unseal` for its tokens
- [ ] Make the scheduler use `token_obj_unseal` for its tokens?

And probably more I forgot.  Presently, this is just a small step in that direction.