@davidchisnall 2023-07-03
Even without the other steps, having a fast-path unseal that avoids cross-compartment calls would be helpful.