@nwf-msr 2023-07-10
This comes with a bit of cost, I'm afraid:

- Baseline (b4b6204e):

  ```
  Test runner: Allocator finished in 4779117 cycles
  Test runner: All tests finished in 5182625 cycles
  ```

- Up through, and including making `token_obj_unseal` a libcall:
  ```
  Test runner: Allocator finished in 4779217 cycles
  Test runner: All tests finished in 5179104 cycles
  ```
  That is, the allocator test is almost completely unchanged (100 instructions slower), but the test suite as a whole got a little bit faster (0.07%), presumably because the relatively few direct calls to `token_obj_unseal` don't go via the compartment switcher now.

- But, once we make the allocator use the `token_obj_unseal` libcall rather than its internal authority,
  ```
  Test runner: Allocator finished in 4803030 cycles
  Test runner: All tests finished in 5203722 cycles
  ```
  The allocator test takes a 0.50% perf hit.  The entire test suite takes a 0.48% hit, since it's a very heavy consumer of `malloc`.

The benefits, of course, are that
- we can now use static ocaps without needing the rest of the heap allocator.
- The stage is set for making the allocator no longer a `PrivilegedCompartment`; it would just need the heap and `MMIO` access to the shadow.  (Once the function pointers are plumbed through, we can also imagine builds with multiple separate allocators.)
- I think we can port the scheduler to the new `token_server` as well, removing some mostly duplicated code from over there.